responsive padding for body
If I was just doing a single theme I could just add this padding to body
and the responsive CSS would override. Since I need to do both depending
on the settings I need to specify a class, and the responsive no longer
overrules it. So need to recreate the logic here.